deplace logique modification vers nouveau controleur
